Create LivePlot_Caffeine_Availability
"""
This is an awesome MVP example of what an auto-updating tkinter-framed graph looks like.

This version accepts new inputs for plotting a different version of the graph - always
starting up with a plot of the last parameters input.

It prints out a statement of what you can expect at 6pm & midnight for caffeine levels

This framework can be used to swap out for any cool plotting you'd like to see over &
over.  Next steps for me are to build out a super quick stats visualizer!  I've left in
the multiple page structure so in the future I can toggle between two different stats
presentations for the same data set.  

Tip:  Not sure why, but after running, you need to close the script, shut down the kernel,
and then reopen it for it to work again :)

To improve this script I would:
--Add in images of an 8, 12, 16, & 20 oz drip coffee cups scaled to their ounceage
--Have it display the last stats on startup on the Label
--Have the ability to add in multiple drinks and add in the values to plot new spikes
 of the caffeine levels as added throughout the day - would look very fancy.
"""